+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved API error responses for some common error cases. (mddburgess/winston#135)
 - Migrated database entities and repositories to Kotlin. (mddburgess/winston#137)
+- Migrated object converters to Kotlin. (mddburgess/winston#138)
 
 ### Fixed